@@ -5,27 +5,29 @@
   "dependencies": {
     "@cfxjs/react-ui": "^2.0.0",
     "@hookform/error-message": "0.0.5",
-    "@hookform/resolvers": "^1.3.2",
+    "@hookform/resolvers": "^1.3.4",
     "big.js": "^6.0.1",
     "classnames": "^2.2.6",
-    "i18next": "^19.7.0",
+    "conflux-address-js": "^1.0.4",
+    "i18next": "^19.8.9",
     "i18next-browser-languagedetector": "^6.0.1",
-    "js-conflux-sdk": "^1.5.6",
+    "perfect-scrollbar": "^1.5.0",
     "qrcode.react": "^1.0.0",
     "react": "^17.0.1",
     "react-copy-to-clipboard": "^5.0.3",
     "react-custom-scrollbars": "^4.2.1",
     "react-dom": "^17.0.1",
-    "react-hook-form": "^6.14.2",
-    "react-i18next": "^11.7.2",
+    "react-hook-form": "^6.15.2",
+    "react-i18next": "^11.8.7",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-transition-group": "^4.4.1",
     "recoil": "^0.1.2",
     "styled-components": "^5.2.1",
+    "swr": "^0.4.2",
     "use-deep-compare-effect": "^1.4.0",
     "wallet-address-validator": "^0.2.4",
-    "yup": "^0.32.8"
+    "yup": "^0.32.9"
   },
   "scripts": {
     "start": "react-app-rewired start",
@@ -55,7 +57,7 @@
     "node-sass": "^4.14.1",
     "prettier": "^2.2.1",
     "react-app-rewired": "^2.1.7",
-    "react-scripts": "^4.0.1",
-    "typescript": "^4.0.3"
+    "react-scripts": "^4.0.2",
+    "typescript": "^4.1.5"
   }
 }
