@@ -3,30 +3,26 @@ import useStyle from '../component/useStyle'
 import Choose from '../token/Choose'
 import styles from './Captain.module.scss'
 import CaptainForm from './FormProvider'
-import { Switch, Route, useRouteMatch } from 'react-router-dom'
+import { useRouteMatch } from 'react-router-dom'
+import useUrlSearch from '../lib/useUrlSearch'
 
-export default function Captain(props) {
-  const {
-    match: { url, path },
-  } = props
+export default function Captain() {
   const [cx] = useStyle(styles)
-  const fullHeight = useRouteMatch({ path, exact: true })
-
+  const { pair } = useUrlSearch()
+  const { url } = useRouteMatch()
+  
   return (
     <MainContainer
       style={{ minHeight: '100%' }}
-      className={fullHeight ? cx('container') : ''}
+      className={!pair ? cx('container') : ''}
     >
-      <Switch>
-        <Route exact path={path}>
-          <div className={cx('container')}>
-            <Choose next={(token) => `${url}/${token}`} captain />
-          </div>
-        </Route>
-        <Route path={`${path}/:erc20`}>
-          <CaptainForm />
-        </Route>
-      </Switch>
+      {pair ? (
+        <CaptainForm pair={pair} />
+      ) : (
+        <div className={cx('container')}>
+          <Choose next={(token) => `${url}?pair=${token}`} captain />
+        </div>
+      )}
     </MainContainer>
   )
 }
