@@ -26,13 +26,14 @@ export default function CaptainForm({
   icon,
   beCaptain,
   cethBalanceBig,
-  burn_fee,
-  mint_fee,
-  symbol,
-  minimal_burn_value,
-  minimal_mint_value,
+  out_fee,
+  in_fee,
+  minimal_out_value,
+  minimal_in_value,
   reference_symbol,
   reference_name,
+  in_token_list,
+  symbol,
   wallet_fee,
   supported,
   sponsor,
@@ -68,12 +69,12 @@ export default function CaptainForm({
   const fields = getFields({
     reference_symbol,
     symbol,
-    mint_fee,
-    burn_fee,
+    out_fee,
+    in_fee,
+    minimal_out_value,
+    minimal_in_value,
     countdown,
     decimals,
-    minimal_burn_value,
-    minimal_mint_value,
     wallet_fee,
     showMortgage,
     cethBalanceBig,
@@ -112,6 +113,7 @@ export default function CaptainForm({
             reference_name,
             supported,
             currentMortgageBig,
+            in_token_list,
             sponsor,
             pendingCount,
             countdown,
