@@ -1,58 +1,43 @@
-import { useState, useMemo, useCallback, useEffect, useRef } from 'react'
+import { useState, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useParams } from 'react-router'
 import useStyle from '../component/useStyle'
 import { buildNum } from '../util/formatNum'
 import useAddress from '../data/useAddress'
-import { useBalance } from '../data/useBalance'
 
-import useTokenList from '../data/useTokenList'
 import useCaptain from '../data/captain'
 
 import formStyles from './Form.module.scss'
 import modalStyles from '../component/modal.module.scss'
-import { CETH_ADDRESS, CONFLUXSCAN_TX } from '../config/config'
+import { CONFLUXSCAN_TX } from '../config/config'
 import createBeCaptain from '../data/beCaptain'
-import getLatestMortgage from '../data/getLatestMortgage'
 import Big from 'big.js'
-import { Loading } from '@cfxjs/react-ui'
 import Modal from '../component/Modal'
 
 import CaptainForm from './Form'
 
 import success from './success.png'
 import fail from './fail.png'
+import useTokenList from '../data/useTokenList'
 
 const MAX_DECIMAL_DISPLAY = 8
 
-//0xd50931bb32fca14acbc0cade5850ba597f3ee1a6
-export default function FormProvider() {
-  const { erc20 } = useParams()
+export default function FormProvider({ pair }) {
   const [popup, setPopup] = useState('')
   const { t } = useTranslation(['captain'])
   const [cx, modalCx] = useStyle(formStyles, modalStyles)
+  const txHash = useRef('')
   const address = useAddress()
-  const cethBalance = useBalance(CETH_ADDRESS)
-  const txHash = useRef()
-
-  /**
-   * tokens will change on every render(no cache in useTokenList)
-   * which will into invalid all the following identity check
-   * no a big problem though
-   */
-  const { tokens } = useTokenList({ erc20 })
-  const tokenInfo = useMemo(
-    () => (tokens && tokens.length > 0 ? tokens[0] : {}),
-    [tokens]
-  )
-
-  const { decimals, sponsor } = tokenInfo
-  const { pendingCount, countdown, minMortgage, cooldownMinutes } = useCaptain(
-    tokenInfo.reference,
-    txHash
-  )
-
-  const [currentMortgage, setCurrentMortgage] = useState()
+  const tokenInfo = useTokenList({ pair })
+  const { decimals, sponsor, reference } = tokenInfo
+  const {
+    pendingCount,
+    countdown,
+    minMortgage,
+    cooldownMinutes,
+    replaceRatio,
+    cethBalance,
+    currentMortgage,
+  } = useCaptain(reference, txHash)
 
   const beCaptain = function ({
     amount,
@@ -64,7 +49,7 @@ export default function FormProvider() {
   }) {
     createBeCaptain(
       address,
-      erc20
+      pair
     )({
       amount: amount && buildNum(amount, 18),
       burnFee: buildNum(burnFee, decimals),
@@ -82,18 +67,6 @@ export default function FormProvider() {
       })
   }
 
-  const updateMinMortgage = useCallback((reference) => {
-    getLatestMortgage(reference).then((x) => {
-      setCurrentMortgage(x && x.toString())
-    })
-  }, [])
-
-  useEffect(() => {
-    if (tokenInfo.reference) {
-      updateMinMortgage(tokenInfo.reference)
-    }
-  }, [updateMinMortgage, tokenInfo.reference])
-
   /**
    * the form default value can be read ONLY ONCE
    * make sure the default from data available when
@@ -106,13 +79,13 @@ export default function FormProvider() {
     cethBalance
   ) {
     const currentMortgageBig = new Big(currentMortgage).div('1e18')
-    let minMortgageBig = new Big(minMortgage).div('1e18')
+    const minMortgageBig = new Big(minMortgage).div('1e18')
 
-    const currentMortgageBigBigger = currentMortgageBig.mul(Big('1.1'))
+    const currentMortgagereplaceBig = currentMortgageBig.mul(replaceRatio)
 
-    if (currentMortgageBigBigger.gt(minMortgageBig)) {
-      minMortgageBig = currentMortgageBigBigger
-    }
+    let minMortgageBigToDisplay = currentMortgagereplaceBig.gt(minMortgageBig)
+      ? currentMortgagereplaceBig
+      : minMortgageBig
 
     const cethBalanceBig = new Big(cethBalance).div('1e18')
 
@@ -121,10 +94,15 @@ export default function FormProvider() {
       cethBalanceDisplay += '...'
     }
 
-    minMortgageBig = minMortgageBig.round(MAX_DECIMAL_DISPLAY, 3)
-    const defaultMortgageBig = sponsor
-      ? minMortgageBig.plus(`1e-${MAX_DECIMAL_DISPLAY}`)
-      : minMortgageBig
+    minMortgageBigToDisplay = minMortgageBigToDisplay.round(
+      MAX_DECIMAL_DISPLAY,
+      3
+    )
+    const defaultMortgageBig =
+      !sponsor || minMortgageBig.gt(minMortgageBigToDisplay)
+        ? minMortgageBig
+        : minMortgageBigToDisplay.plus(`1e-${MAX_DECIMAL_DISPLAY}`)
+
     const data = {
       address,
       ...tokenInfo,
@@ -135,7 +113,7 @@ export default function FormProvider() {
 
       beCaptain,
       minMortgage,
-      minMortgageBig,
+      minMortgageBig: minMortgageBigToDisplay,
       currentMortgageBig,
       defaultMortgageBig,
       cethBalanceBig,
@@ -172,12 +150,6 @@ export default function FormProvider() {
           )}
         </Modal>
       </>
-    )
-  } else {
-    return (
-      <div className={cx('loading-container')}>
-        <Loading size="large" />
-      </div>
     )
   }
 }
