@@ -7,12 +7,13 @@ const basicValidate = () => {
 export default function getFields({
   reference_symbol,
   symbol,
-  mint_fee,
-  burn_fee,
+
+  out_fee,
+  in_fee,
+  minimal_out_value,
+  minimal_in_value,
   countdown,
   decimals,
-  minimal_burn_value,
-  minimal_mint_value,
   wallet_fee,
   showMortgage,
   cethBalanceBig,
@@ -26,14 +27,14 @@ export default function getFields({
       unit: symbol,
       decimals,
       validate: basicValidate(),
-      defaultValue: mint_fee,
+      defaultValue: in_fee,
       readOnly: countdown !== 0,
     },
     {
       label: 'shuttle-in-amount',
       unit: reference_symbol,
       name: 'minimal_mint_value',
-      defaultValue: minimal_mint_value,
+      defaultValue: minimal_in_value,
       validate: basicValidate().greaterThan('mint_fee', 'error.above-in-fee'),
       decimals,
       readOnly: countdown !== 0,
@@ -42,7 +43,7 @@ export default function getFields({
       label: 'shuttle-out-fee',
       unit: symbol,
       name: 'burn_fee',
-      defaultValue: burn_fee,
+      defaultValue: out_fee,
       validate: basicValidate(),
       decimals,
       readOnly: countdown !== 0,
@@ -51,7 +52,7 @@ export default function getFields({
       label: 'shuttle-out-amount',
       unit: symbol,
       name: 'minimal_burn_value',
-      defaultValue: minimal_burn_value,
+      defaultValue: minimal_out_value,
       validate: basicValidate().greaterThan('burn_fee', 'error.above-out-fee'),
       decimals,
       readOnly: countdown !== 0,
