@@ -1,4 +1,5 @@
-export const breakPoint = 675
+// export const breakPoint = 1100
+export const breakPoint = 800
 
 export const device = {
   mobile: `(max-width: ${breakPoint}px)`,
