@@ -1,92 +1,72 @@
-import { useEffect } from 'react'
-import useState1 from './useState1'
+import { useParams } from 'react-router'
+import CHAIN_CONFIG from '../config/chainConfig'
 
-import tokenList from './tokenList'
-import { isAddress } from '../util/address'
-import jsonrpc from './jsonrpc'
+import { getTokenList } from './tokenList'
+import useSWR from 'swr'
+import { CHAIN_SINGLE_PAIR } from '../config/constant'
 
-let supportedTokensResolved
-
-
-export default function useTokenList({ search, erc20 = '', cToken } = {}) {
-  const [state, setState] = useState1({
-    tokens: [],
-    isLoading: supportedTokensResolved,
-  })
-  if (isAddress(search)) {
-    if (!cToken) {
-      erc20 = search
-    }
+function fetcher(key, searchOrPair, chain, cToken) {
+  let search, pair
+  if (key === 'search') {
+    search = searchOrPair
+  } else if (key === 'pair') {
+    pair = searchOrPair
   }
+  const { singleToken, display,searchList } = CHAIN_CONFIG[chain]
+  return getTokenList(chain).then(({ tokenList, tokenMap }) => {
+    if (pair) {
+      return singleToken
+        ? { ...tokenList[0], singleton: true }
+        : pair === CHAIN_SINGLE_PAIR
+        ? null
+        : tokenMap[pair]
+    }
+    if (!search) {
+      return tokenList.filter(display)
+    }
+    return (cToken ? searchCfxList : searchList)(tokenList, search).then((list) =>
+      sortSearchResult(list)
+    )
+  })
+}
+export default function useTokenList({ pair, search, cToken } = {}) {
+  const { chain } = useParams()
+  return useSWR(
+    pair ? ['pair', pair, chain] : ['search', search, chain, cToken],
+    fetcher,
+    { suspense: true }
+  ).data
+}
 
-  erc20 = erc20.toLocaleLowerCase()
-  useEffect(() => {
-    return tokenList
-      .then((tokens) => {
-        if (erc20) {
-          return Promise.resolve(
-            tokens.filter(({ reference }) => erc20 === reference)
-          ).then(([token]) => {
-            if (!token) {
-              return jsonrpc('searchToken', {
-                url: 'sponsor',
-                params: [erc20 || search],
-              }).then((result) => {
-                if (result && result.is_valid_erc20) {
-                  return [result]
-                } else {
-                  return []
-                }
-              })
-            } else {
-              return [token]
-            }
-          })
-        } else if (search) {
-          const lowersearch = search.toLowerCase()
-          if (isAddress(search)) {
-            //MUST be ctoken search
-            return tokens.filter(({ ctoken }) => lowersearch === ctoken)
-          } else {
-            return tokens
-              .filter(
-                ({ reference_symbol, reference_name, supported, symbol }) => {
-                  return (
-                    (reference_symbol.toLowerCase().indexOf(lowersearch) > -1 ||
-                      symbol.toLowerCase().indexOf(lowersearch) > -1 ||
-                      reference_name.toLowerCase().indexOf(lowersearch) > -1) &&
-                    (cToken ? supported === 1 : true)
-                  )
-                }
-              )
-              .slice()
-              .sort(
-                (
-                  {
-                    supported: supported0,
-                    in_token_list: in_token_list0,
-                  },
-                  {
-                    supported: supported1,
-                    in_token_list: in_token_list1,
-                  }
-                ) => {
-                  return (
-                    supported1 - supported0 || in_token_list1 - in_token_list0
-                  )
-                }
-              )
-          }
-        } else {
-          return tokens.filter(
-            ({ supported, in_token_list }) =>
-              supported === 1 && in_token_list === 1
-          )
-        }
-      })
-      .then((tokens) => {
-        setState({ tokens: tokens.filter((x) => x), isLoading: false })
-      })
-  }, [search, setState, erc20, cToken])
-  return state
+
+function searchCfxList(list, search) {
+  const lowerSearch = search.toLowerCase()
+  return Promise.resolve(
+    list.filter(
+      ({ reference_name, reference_symbol, ctoken, symbol, supported }) => {
+        return (
+          //DO NOT present unsupported with ctoken
+          supported &&
+          (ctoken === search ||
+            reference_symbol.toLowerCase().indexOf(lowerSearch) > -1 ||
+            symbol.toLowerCase().indexOf(lowerSearch) > -1 ||
+            reference_name.toLowerCase().indexOf(lowerSearch) > -1)
+        )
+      }
+    )
+  )
+}
+
+
+function sortSearchResult(list) {
+  return list
+    .slice()
+    .sort(
+      (
+        { supported: supported0, in_token_list: in_token_list0 },
+        { supported: supported1, in_token_list: in_token_list1 }
+      ) => {
+        return supported1 - supported0 || in_token_list1 - in_token_list0
+      }
+    )
 }
