@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useState } from 'react'
 import { Switch, Route, Link, Redirect, useParams } from 'react-router-dom'
 import { useRecoilState } from 'recoil'
 import layoutBottomState from '../layout/LayoutButtomState'
@@ -18,9 +18,12 @@ import MenuLink from '../component/MenuLink'
 import PaddingContainer from '../component/PaddingContainer/PaddingContainer'
 import MainContainer from '../component/MainContainer/MainContainer'
 import useStyle from '../component/useStyle'
-import useTokenList from '../data/useTokenList'
 import Spec from '../layout/Spec'
 import useIsSamll from '../component/useSmallScreen'
+import useUrlSearch from '../lib/useUrlSearch'
+import useTokenList from '../data/useTokenList'
+import { CHAIN_SINGLE_PAIR } from '../config/constant'
+import Modal from '../component/Modal'
 export default function Shuttle({ match: { path, url } }) {
   const [cx] = useStyle(styles)
   const { t } = useTranslation(['nav'])
@@ -75,10 +78,7 @@ export default function Shuttle({ match: { path, url } }) {
       <PaddingContainer bottom>
         <Switch>
           <Redirect from={path} exact to={`${path}/in`} />
-          <Route
-            path={`${path}/:type/:erc20?`}
-            component={RouteComponent}
-          ></Route>
+          <Route path={`${path}/:type`} component={RouteComponent}></Route>
         </Switch>
       </PaddingContainer>
     </MainContainer>
@@ -86,10 +86,16 @@ export default function Shuttle({ match: { path, url } }) {
 }
 
 function RouteComponent() {
-  const { type, erc20 = '' } = useParams()
-  const { tokens } = useTokenList({ erc20 })
-  //display tokenInfo only when token is url available
-  const tokenInfo = erc20 && tokens ? tokens[0] : null
+  const { pair = '' } = useUrlSearch()
+  const tokenInfo = useTokenList({ pair: pair || CHAIN_SINGLE_PAIR })
+  const [feePopup, setFeePopup] = useState(false)
+  const { type } = useParams()
+
   const Component = type === 'in' ? ShuttleIn : ShuttleOut
-  return <Component tokenInfo={tokenInfo} />
+
+  return (
+    <>
+      <Component tokenInfo={tokenInfo} {...{ feePopup, setFeePopup }} />
+    </>
+  )
 }
