@@ -2,27 +2,28 @@ import React, { useCallback, useState } from 'react'
 
 import { useTranslation, Trans } from 'react-i18next'
 import QRCode from 'qrcode.react'
+import { CopyToClipboard } from 'react-copy-to-clipboard'
 
 import down from '../down.svg'
 import copy from './i-copy-48.png'
 
 import tick from './tick.svg'
 import qr from './qr.svg'
-import { CopyToClipboard } from 'react-copy-to-clipboard'
 
-import Modal from '../../component/Modal'
-import modalStyles from '../../component/modal.module.scss'
+import Modal, { modalStyles } from '../../component/Modal'
+
 import ShuttleHistory from '../../history/ShuttleHistory'
 
 import useStyle from '../../component/useStyle'
 import commonInputStyles from '../../component/input.module.scss'
 import shuttleStyle from '../Shuttle.module.scss'
 import shuttleInStyles from './ShuttleIn.module.scss'
-import useShuttleInAddress from '../../data/useShuttleInAddress'
+import useShuttleAddress from '../../data/useShuttleInAddress'
 import TokenInput from '../TokenInput'
 
-import CTokenPopup from '../CTokenPopup'
 import WithQuestion from '../../component/WithQuestion'
+import { Loading } from '@cfxjs/react-ui'
+import { useParams } from 'react-router'
 
 export default function ShuttleIn({ tokenInfo }) {
   const [commonCx, shuttleCx, shuttleInCx, modalCx] = useStyle(
@@ -31,15 +32,11 @@ export default function ShuttleIn({ tokenInfo }) {
     shuttleInStyles,
     modalStyles
   )
-  const { t } = useTranslation('shuttle-in')
-  const shuttleInAddress = useShuttleInAddress(tokenInfo)
-
+  const { t } = useTranslation(['shuttle-in', 'shuttle'])
   const [addressPopup, setAddressPopup] = useState(false)
-  const [cTokenPopup, setCTokenPopup] = useState(false)
   const [minPopup, setMinPopup] = useState(false)
-  const [feePopup, setFeePopup] = useState(false)
+
   const [copyPopup, setCopyPopup] = useState(false)
-  const [qrPopup, setQrPopup] = useState(false)
 
   const displayCopy = useCallback(() => {
     setCopyPopup(true)
@@ -52,87 +49,38 @@ export default function ShuttleIn({ tokenInfo }) {
   return (
     <div className={shuttleInCx('container')}>
       <TokenInput
-        to={{
-          pathname: '/token',
-          search: '?next=/shuttle/in',
-        }}
         tokenInfo={tokenInfo}
+        dir="from"
         placeholder={t('placeholder.out')}
+        displayCopy={displayCopy}
       />
       <div className={shuttleCx('down')}>
         <img alt="down" src={down}></img>
       </div>
       <TokenInput
-        to={{
-          pathname: '/token',
-          search: `?next=/shuttle/in&cToken=1`,
-        }}
+        dir="to"
         tokenInfo={tokenInfo}
         placeholder={t('placeholder.in')}
-        cToken={() => setCTokenPopup(true)}
+        displayCopy={displayCopy}
+        cToken
       />
 
       {tokenInfo && (
-        <div className={shuttleCx('small-text')}>
-          <WithQuestion onClick={() => setMinPopup(true)}>
-            <span>{t('amount', tokenInfo)}</span>
-          </WithQuestion>
-
-          <WithQuestion onClick={() => setFeePopup(true)}>
-            <span>{t('fee', tokenInfo)}</span>
-          </WithQuestion>
-        </div>
+        <TokenInfoDetails
+          {...{
+            shuttleCx,
+            setMinPopup,
+            tokenInfo,
+            shuttleInCx,
+            t,
+            modalCx,
+            commonCx,
+            setAddressPopup,
+            displayCopy,
+          }}
+        />
       )}
-
-      <div className={shuttleInCx('address')}>
-        <WithQuestion
-          className={shuttleCx('title')}
-          onClick={() => setAddressPopup(true)}
-        >
-          <span>{t('address')}</span>
-        </WithQuestion>
-
-        <div className={shuttleInCx('address-input')}>
-          <input
-            readOnly
-            defaultValue={shuttleInAddress}
-            className={
-              commonCx('input-common') + ' ' + shuttleInCx('input-address')
-            }
-            placeholder={t('address-placeholder')}
-          />
-          {shuttleInAddress && (
-            <CopyToClipboard text={shuttleInAddress} onCopy={displayCopy}>
-              <img alt="copy" className={shuttleInCx('copy')} src={copy}></img>
-            </CopyToClipboard>
-          )}
-        </div>
-      </div>
-      {tokenInfo && (
-        <div
-          style={{ alignItems: 'flex-start' }}
-          className={shuttleCx('small-text')}
-        >
-          <span>
-            <Trans
-              t={t}
-              i18nKey="latest"
-              values={{
-                type: t(tokenInfo.reference === 'btc' ? 'btc' : 'eth'),
-              }}
-            ></Trans>
-          </span>
-
-          <span
-            className={shuttleInCx('qr-container')}
-            onClick={() => setQrPopup(true)}
-          >
-            <img className={shuttleInCx('img')} alt="qr" src={qr}></img>
-            <span>{t('qr')}</span>
-          </span>
-        </div>
-      )}
-      <ShuttleHistory type="mint" />
+      <ShuttleHistory type="in" />
       <Modal
         show={addressPopup}
         title
@@ -141,25 +89,6 @@ export default function ShuttleIn({ tokenInfo }) {
       >
         <div className={modalCx('content')}>{t('popup.address')}</div>
         <div className={modalCx('btn')} onClick={() => setAddressPopup(false)}>
-          {t('popup.ok')}
-        </div>
-      </Modal>
-      <CTokenPopup
-        displayCopy={displayCopy}
-        cTokenPopup={cTokenPopup}
-        setCTokenPopup={setCTokenPopup}
-        tokenInfo={tokenInfo}
-      />
-      <Modal
-        show={feePopup}
-        title
-        onClose={() => setFeePopup(false)}
-        clickAway={() => setFeePopup(false)}
-      >
-        <div className={modalCx('content')}>
-          <Trans values={tokenInfo} t={t} i18nKey="popup.fee"></Trans>
-        </div>
-        <div className={modalCx('btn')} onClick={() => setFeePopup(false)}>
           {t('popup.ok')}
         </div>
       </Modal>
@@ -185,6 +114,88 @@ export default function ShuttleIn({ tokenInfo }) {
           <div>{t('popup.copy')}</div>
         </div>
       </Modal>
+    </div>
+  )
+}
+
+function TokenInfoDetails({
+  shuttleCx,
+  commonCx,
+  modalCx,
+  shuttleInCx,
+  t,
+  setMinPopup,
+  setAddressPopup,
+  displayCopy,
+  tokenInfo,
+}) {
+  const { origin } = tokenInfo
+  const { chain } = useParams()
+  const [feePopup, setFeePopup] = useState(false)
+  const [qrPopup, setQrPopup] = useState(false)
+  const shuttleInAddress = useShuttleAddress({ type: 'in', origin })
+  return (
+    <>
+      <div className={shuttleCx('small-text')}>
+        <WithQuestion onClick={() => setMinPopup(true)}>
+          <span>{t('amount', tokenInfo)}</span>
+        </WithQuestion>
+
+        <WithQuestion onClick={() => setFeePopup(true)}>
+          <span>{t('fee', tokenInfo)}</span>
+        </WithQuestion>
+      </div>
+
+      <div className={shuttleInCx('address')}>
+        <WithQuestion
+          className={shuttleCx('title')}
+          onClick={() => setAddressPopup(true)}
+        >
+          <span>{t('address')}</span>
+        </WithQuestion>
+
+        <div className={shuttleInCx('address-input')}>
+          <input
+            readOnly
+            defaultValue={shuttleInAddress}
+            className={
+              commonCx('input-common') + ' ' + shuttleInCx('input-address')
+            }
+            placeholder={t('address-placeholder')}
+          />
+          {shuttleInAddress ? (
+            <CopyToClipboard text={shuttleInAddress} onCopy={displayCopy}>
+              <img alt="copy" className={shuttleInCx('copy')} src={copy}></img>
+            </CopyToClipboard>
+          ) : (
+            <div className={shuttleInCx('copy')}>
+              <Loading size="1rem" />
+            </div>
+          )}
+        </div>
+      </div>
+      <div
+        style={{ alignItems: 'flex-start' }}
+        className={shuttleCx('small-text')}
+      >
+        <span>
+          <Trans
+            t={t}
+            i18nKey="latest"
+            values={{
+              type: t(chain),
+            }}
+          ></Trans>
+        </span>
+
+        <span
+          className={shuttleInCx('qr-container')}
+          onClick={() => setQrPopup(true)}
+        >
+          <img className={shuttleInCx('img')} alt="qr" src={qr}></img>
+          <span>{t('qr')}</span>
+        </span>
+      </div>
 
       <Modal
         onClose={() => setQrPopup(false)}
@@ -206,6 +217,29 @@ export default function ShuttleIn({ tokenInfo }) {
           </div>
         </div>
       </Modal>
-    </div>
+      <Modal
+        show={feePopup}
+        title
+        onClose={() => setFeePopup(false)}
+        clickAway={() => setFeePopup(false)}
+      >
+        <div className={modalCx('content')}>
+          <Trans
+            values={{
+              operation: t('shuttle-in'),
+              ...tokenInfo,
+              fee: tokenInfo['in_fee'],
+              is_create:
+                origin === chain ? t('wallet-create-fee', tokenInfo) : '',
+            }}
+            t={t}
+            i18nKey="popup-fee"
+          ></Trans>
+        </div>
+        <div className={modalCx('btn')} onClick={() => setFeePopup(false)}>
+          {t('popup.ok')}
+        </div>
+      </Modal>
+    </>
   )
 }
